+++ sdk/communication/azure-communication-sms/src/test/java/com/azure/communication/sms/SmsAsyncClientTests.java
@@ -15,6 +15,8 @@ import reactor.test.StepVerifier;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class SmsAsyncClientTests extends SmsTestBase {
@@ -29,13 +31,12 @@ public class SmsAsyncClientTests extends SmsTestBase {
     @ParameterizedTest
     @MethodSource("com.azure.core.test.TestBase#getHttpClients")
     public void createAsyncClientUsingConnectionString(HttpClient httpClient) {
-
         to = new ArrayList<String>();
-        to.add(TO_PHONE_NUMBER);
+        to.add(SMS_SERVICE_PHONE_NUMBER);
         SmsClientBuilder builder = getSmsClientUsingConnectionString(httpClient);
         asyncClient = setupAsyncClient(builder, "createAsyncSmsClientUsingConnectionString");
         assertNotNull(asyncClient);
-        Mono<Iterable<SmsSendResult>> response = asyncClient.send(FROM_PHONE_NUMBER, to, MESSAGE, null);
+        Mono<Iterable<SmsSendResult>> response = asyncClient.send(SMS_SERVICE_PHONE_NUMBER, to, MESSAGE, null);
         StepVerifier.create(response)
             .assertNext(item -> {
                 assertNotNull(item);
@@ -50,11 +51,11 @@ public class SmsAsyncClientTests extends SmsTestBase {
     public void sendSmsUsingTokenCredential(HttpClient httpClient) {
         TokenCredential tokenCredential = new DefaultAzureCredentialBuilder().build();
         to = new ArrayList<>();
-        to.add(TO_PHONE_NUMBER);
+        to.add(SMS_SERVICE_PHONE_NUMBER);
         SmsClientBuilder  builder = getSmsClientWithToken(httpClient, tokenCredential);
-        asyncClient = setupAsyncClient(builder, "createAsyncSmsClientUsingConnectionString");
+        asyncClient = setupAsyncClient(builder, "sendSmsUsingTokenCredential");
         assertNotNull(asyncClient);
-        Mono<Iterable<SmsSendResult>> response = asyncClient.send(FROM_PHONE_NUMBER, to, MESSAGE, null);
+        Mono<Iterable<SmsSendResult>> response = asyncClient.send(SMS_SERVICE_PHONE_NUMBER, to, MESSAGE, null);
         StepVerifier.create(response)
             .assertNext(item -> {
                 assertNotNull(item);
@@ -66,16 +67,15 @@ public class SmsAsyncClientTests extends SmsTestBase {
     @ParameterizedTest
     @MethodSource("com.azure.core.test.TestBase#getHttpClients")
     public void sendToMultipleUsers(HttpClient httpClient) {
-
         to = new ArrayList<String>();
-        to.add(TO_PHONE_NUMBER);
+        to.add(SMS_SERVICE_PHONE_NUMBER);
         // Arrange
         SmsClientBuilder builder = getSmsClient(httpClient);
-        asyncClient = setupAsyncClient(builder, "send");
+        asyncClient = setupAsyncClient(builder, "sendToMultipleUsers");
 
         // Action & Assert
 
-        Mono<Iterable<SmsSendResult>> response = asyncClient.send(FROM_PHONE_NUMBER, to, MESSAGE, null);
+        Mono<Iterable<SmsSendResult>> response = asyncClient.send(SMS_SERVICE_PHONE_NUMBER, to, MESSAGE, null);
         StepVerifier.create(response)
             .assertNext(item -> {
                 assertNotNull(item);
@@ -86,15 +86,13 @@ public class SmsAsyncClientTests extends SmsTestBase {
     @ParameterizedTest
     @MethodSource("com.azure.core.test.TestBase#getHttpClients")
     public void sendToSingleUser(HttpClient httpClient) {
-
-
         // Arrange
         SmsClientBuilder builder = getSmsClient(httpClient);
-        asyncClient = setupAsyncClient(builder, "send");
+        asyncClient = setupAsyncClient(builder, "sendToSingleUser");
 
         // Action & Assert
 
-        Mono<SmsSendResult> response = asyncClient.send(FROM_PHONE_NUMBER, TO_PHONE_NUMBER, MESSAGE, null);
+        Mono<SmsSendResult> response = asyncClient.send(SMS_SERVICE_PHONE_NUMBER, SMS_SERVICE_PHONE_NUMBER, MESSAGE, null);
         StepVerifier.create(response)
             .assertNext(item -> {
                 assertNotNull(item);
@@ -110,14 +108,13 @@ public class SmsAsyncClientTests extends SmsTestBase {
         options.setDeliveryReportEnabled(true);
         options.setTag("New Tag");
 
-
         // Arrange
         SmsClientBuilder builder = getSmsClient(httpClient);
-        asyncClient = setupAsyncClient(builder, "send");
+        asyncClient = setupAsyncClient(builder, "sendToSingleUserWithOptions");
 
         // Action & Assert
 
-        Mono<SmsSendResult> response = asyncClient.send(FROM_PHONE_NUMBER, TO_PHONE_NUMBER, MESSAGE, options);
+        Mono<SmsSendResult> response = asyncClient.send(SMS_SERVICE_PHONE_NUMBER, SMS_SERVICE_PHONE_NUMBER, MESSAGE, options);
         StepVerifier.create(response)
             .assertNext(item -> {
                 assertNotNull(item);
@@ -128,20 +125,17 @@ public class SmsAsyncClientTests extends SmsTestBase {
     @ParameterizedTest
     @MethodSource("com.azure.core.test.TestBase#getHttpClients")
     public void sendToMultipleUsersWithOptions(HttpClient httpClient) {
-
         SmsSendOptions options = new SmsSendOptions();
         options.setDeliveryReportEnabled(true);
         options.setTag("New Tag");
-
         to = new ArrayList<String>();
-        to.add(TO_PHONE_NUMBER);
+        to.add(SMS_SERVICE_PHONE_NUMBER);
         // Arrange
         SmsClientBuilder builder = getSmsClient(httpClient);
-        asyncClient = setupAsyncClient(builder, "send");
+        asyncClient = setupAsyncClient(builder, "sendToMultipleUsersWithOptions");
 
         // Action & Assert
-
-        Mono<Iterable<SmsSendResult>> response = asyncClient.send(FROM_PHONE_NUMBER, to, MESSAGE, options);
+        Mono<Iterable<SmsSendResult>> response = asyncClient.send(SMS_SERVICE_PHONE_NUMBER, to, MESSAGE, options);
         StepVerifier.create(response)
             .assertNext(item -> {
                 assertNotNull(item);
@@ -152,25 +146,24 @@ public class SmsAsyncClientTests extends SmsTestBase {
     @ParameterizedTest
     @MethodSource("com.azure.core.test.TestBase#getHttpClients")
     public void sendToIncorrectPhoneNumber(HttpClient httpClient) {
-
         to = new ArrayList<String>();
-        to.add(FAIL_PHONE_NUMBER);
+        to.add("+155512345678");
         SmsClientBuilder builder = getSmsClientUsingConnectionString(httpClient);
-        asyncClient = setupAsyncClient(builder, "createAsyncSmsClientUsingConnectionString");
+        asyncClient = setupAsyncClient(builder, "sendToIncorrectPhoneNumber");
         assertNotNull(asyncClient);
-        Mono<Iterable<SmsSendResult>> response = asyncClient.send(FROM_PHONE_NUMBER, to, MESSAGE, null);
+        Mono<Iterable<SmsSendResult>> response = asyncClient.send(SMS_SERVICE_PHONE_NUMBER, to, MESSAGE, null);
         assertNotNull(response);
         StepVerifier.create(response)
             .assertNext(item -> {
                 assertNotNull(item);
             })
             .verifyComplete();
+        Iterable<SmsSendResult> smsSendResults = response.block();
+        for (SmsSendResult result : smsSendResults) {
+            assertFalse(result.isSuccessful());
+        }
+
     }
-
-
-
-
-
 
     private SmsAsyncClient setupAsyncClient(SmsClientBuilder builder, String testName) {
         return addLoggingPolicy(builder, testName).buildAsyncClient();